Move metrics tracker from reporting to shared utils
Relocate src/shared/reporting/metrics-tracker.js into src/shared/utils/metrics-tracker.js so the telemetry helper sits alongside other shared utilities, update barrels/imports to use the utils entry point, and keep the reporting shim forwarding to the new path.
