+++ src/shared/reporting/index.js
@@ -1,1 +1,1 @@
-export * from "./metrics-tracker.js";
+export * from "../utils/metrics-tracker.js";
