+++ src/shared/utils/index.js
@@ -4,6 +4,7 @@ export * from "./capability-probes.js";
 export * from "./error.js";
 export * from "./json.js";
 export * from "./line-breaks.js";
+export * from "./metrics-tracker.js";
 export * from "./enumerated-options.js";
 export * from "./number.js";
 export * from "./numeric-options.js";
